updates to basic_project_stats.py

+++ example_scripts/basic_project_stats.py
@@ -8,8 +8,8 @@ try:
     classfile_in = sys.argv[1]
 except:
     print("\nUsage: %s classifications_infile" % sys.argv[0])
-    print("      classifications_infile is a Zooniverse (Panoptes) classifications data export CSV.")
-    print("  Optional inputs (no spaces):")
+    print("      classifications_infile is a Zooniverse (Panoptes) classifications data export CSV.\n")
+    print("  Optional inputs:")
     print("    workflow_id=N")
     print("       specify the program should only consider classifications from workflow id N")
     print("    workflow_version=M")
@@ -22,7 +22,11 @@ except:
     print("    --remove_duplicates")
     print("       remove duplicate classifications (subject-user pairs) before analysis.")
     print("       memory-intensive for big files; probably best to pair with outfile_csv so you save the output.")
-    print("\nAll output will be to stdout (about a paragraph worth).\n")
+    print("    --keep_nonlive")
+    print("       by default the program ignores classifications made while the project wasn't 'Live'; setting this will keep them in.")
+    print("    --keep_allcols")
+    print("       by default the program only keeps columns required for stats; use this with a specified outfile_csv to save all columns, including annotations. (If you're not using outfile_csv this will just waste memory.)")
+    print("\nAll output will be to stdout (about 1-2 paragraphs' worth).\n")
     sys.exit(0)
 
 
@@ -31,7 +35,7 @@ import numpy as np  # works in 1.10.1
 import pandas as pd  # works in 0.13.1
 import datetime
 import dateutil.parser
-import json
+import json, ujson
 import gc
 
 # default value is not to care about workflow ID or version
@@ -46,6 +50,10 @@ output_csv = False
 # the Zooniverse has squashed several bugs related to this, but some still
 # happen client-side and there's nothing we can do about that.
 remove_duplicates = False
+# by default, restrict the analysis to "Live" classifications
+keep_nonlive = False
+# by default, don't keep every column of the classifications when writing to an outfile
+keep_allcols = False
 
 # check for other command-line arguments
 if len(sys.argv) > 2:
@@ -57,13 +65,17 @@ if len(sys.argv) > 2:
             workflow_id = int(arg[1])
         elif arg[0] == "workflow_version":
             workflow_version = float(arg[1])
-        elif arg[0] == "outfile_csv":
+        elif (arg[0] == "outfile_csv") | (arg[0] == "outfile"):
             outfile_csv = arg[1]
             output_csv  = True
+        elif arg[0] == "--keep_allcols":
+            keep_allcols = True
         elif arg[0] == "--time_elapsed":
             time_elapsed = True
         elif arg[0] == "--remove_duplicates":
             remove_duplicates = True
+        elif arg[0] == "--keep_nonlive":
+            keep_nonlive = True
 
 
 
@@ -156,7 +168,8 @@ def gini(list_of_values):
 
 def get_duplicate_ids(grp):
     # groupbys and dfs have slightly different indexing and just NOPE
-    thegrp = pd.DataFrame(grp)
+    #thegrp = pd.DataFrame(grp)
+    thegrp = grp
 
     if len(thegrp) == 1:
         return
@@ -166,6 +179,19 @@ def get_duplicate_ids(grp):
 
 
 
+
+def get_live_project(meta_json):
+    try:
+        return meta_json['live_project']
+    except:
+        # apparently some subject metadata doesn't have this? dunno?
+        return False
+
+def get_live_project_incl_missing(meta_json):
+    try:
+        return meta_json['live_project']
+    except:
+        return -1
 
 # Begin the main stuff
 
@@ -186,7 +212,22 @@ big projects with big catalogs it can take up a lot of memory, so we don't want 
 use it if we don't have to.
 '''
 cols_keep = ["classification_id", "user_name", "user_id", "user_ip", "workflow_id", "workflow_version", "created_at", "metadata", "subject_ids"]
-classifications = pd.read_csv(classfile_in, usecols=cols_keep)
+if not keep_allcols:
+    try:
+        classifications = pd.read_csv(classfile_in, usecols=cols_keep)
+    except:
+        print("Some columns missing from classifications infile, reading without specifying columns (uses more memory)... ")
+        classifications = pd.read_csv(classfile_in)
+else:
+    classifications = pd.read_csv(classfile_in, low_memory=False)
+    cols_used = classifications.columns.tolist()
+    cols_out  = classifications.columns.tolist()
+    if not 'created_day' in cols_used:
+        cols_used.append('created_day')
+    if not 'meta_json' in cols_used:
+        cols_used.append('meta_json')
+
+n_class_raw = len(classifications)
 
 # now restrict classifications to a particular workflow id/version if requested
 if (workflow_id > 0) | (workflow_version > 0):
@@ -237,15 +278,52 @@ else:
     print(" and workflow_versions:")
     print(version_ints)
 
+
+# Remove classifications collected before the project went Live
+# note: it makes logical sense to do this *before* we extract the classifications
+# from the workflow we care about, *but* the meta_json setting step (which we
+# need in order to extract Live project status) can take a while (up to ~minutes)
+# and adds to memory usage, so I'd rather do it after we've already culled
+# the table of potentially a lot of unused rows.
+# OTOH culling duplicates takes more time and memory than culling unused workflow
+# versions, so wait to do that until after we've removed non-Live classifications
+
+# first, extract the metadata column into a json we can read entries for
+#
+# ujson is quite a bit faster than json but seems to use a bit more memory as it works
+classifications['meta_json'] = [ujson.loads(q) for q in classifications.metadata]
+
+if keep_nonlive:
+    print("Retaining all non-live classifications in analysis.")
+else:
+    # would that we could just do q['live_project'] but if that tag is missing for
+    # any classifications (which it is in some cases) it crashes
+    classifications['live_project']  = [get_live_project(q) for q in classifications.meta_json]
+
+    # if this line gives you an error you've read in this boolean as a string
+    # so need to convert "True" --> True and "False" --> False
+    class_live = classifications[classifications.live_project].copy()
+    n_class_thiswf = len(classifications)
+    n_live = sum(classifications.live_project)
+    n_notlive = n_class_thiswf - n_live
+    print(" Removing %d non-live classifications..." % n_notlive)
+
+    # don't make a slice but also save memory
+    classifications = pd.DataFrame(class_live)
+    del class_live
+    gc.collect()
+
+
+
 # if we've been asked to remove duplicates, do that now
 if remove_duplicates:
     '''
     a duplicate can be that the classification id is submitted twice by the client
     but it can also be that the classifier classified the same subject twice in different classification_ids.
 
-    So identify duplicates based on username and subject id, not based on classification_id.
+    So identify duplicates based on username + subject id + workflow info, not based on classification_id.
     '''
-    subj_classifications = classifications.groupby('user_name subject_ids'.split())
+    subj_classifications = classifications.groupby('user_name subject_ids workflow_id workflow_version'.split())
 
     n_class = len(classifications)
     # just take the first of each of the groups
@@ -264,26 +342,61 @@ if remove_duplicates:
         print("Found %d duplicate classifications (%.2f percent of total)." % (n_dups, float(n_dups)/float(n_class)*100.0))
 
         # get the duplicate classifications and save them before we remove them
-        class_dups = pd.DataFrame(subj_classifications.apply(get_duplicate_ids))
-        is_a_dup = [np.invert(np.isnan(q)) for q in class_dups.workflow_id]
-        class_dups = class_dups[is_a_dup]
+        #class_dups = pd.DataFrame(subj_classifications.apply(get_duplicate_ids))
+
+        # if you want to keep a record of everything with just the dups flagged,
+        # this is your thing
+        #dups_flagged = pd.merge(classifications, classifications_nodups['classification_id subject_id'.split()], how='outer', on='classification_id', suffixes=('', '_2'), indicator=True)
+        # if you just need something that has only the dups in it, here you go
+        dups_only = classifications[~classifications.isin(classifications_nodups)].dropna(how='all')
+
+        # dups_only has the duplicates only - not the original classification in each set
+        # i.e. if classifications 123, 456, and 789 are all from the same user
+        # classifying the same subject, dups_only will only contain classifications
+        # 456 and 789. When we save the duplicate classifications we want to save
+        # the initial classification (that was later duplicated) as well, so we
+        # need to retrieve those.
+        # I don't see a really easy way to do it based on the groupby we already did
+        # (subj_classifications)
+        # so let's just define what identifies the duplicate (user_name + subject_ids)
+        # and pick them out.
+        # even for a reasonably big dataset this is relatively fast (seconds, not minutes)
+        try:
+            dups_only['user_subj_pair'] = dups_only['user_name']+'_'+dups_only['subject_ids'].astype(int).astype(str)+'_'+dups_only['workflow_id'].astype(str)+'v'+dups_only['workflow_version'].astype(str)
+        except:
+            dups_only['user_subj_pair'] = dups_only['user_name']+'_'+dups_only['subject_ids'].astype(str)+'_'+dups_only['workflow_id'].astype(str)+'v'+dups_only['workflow_version'].astype(str)
+
+        # n_dup_pairs tracks unique user-subject pairs that were duplicated
+        dup_pairs = dups_only['user_subj_pair'].unique()
+        n_dup_pairs = len(dup_pairs)
+
+        try:
+            classifications['user_subj_pair'] = classifications['user_name']+'_'+classifications['subject_ids'].astype(int).astype(str)+'_'+classifications['workflow_id'].astype(str)+'v'+classifications['workflow_version'].astype(str)
+        except:
+            classifications['user_subj_pair'] = classifications['user_name']+'_'+classifications['subject_ids'].astype(str)+'_'+classifications['workflow_id'].astype(str)+'v'+classifications['workflow_version'].astype(str)
+
+        # this keeps things that are any part of a duplicate set, including first
+        is_a_dup = classifications['user_subj_pair'].isin(dup_pairs)
+
+        class_dups = classifications[is_a_dup].copy()
+        # counts any classification that is any part of a duplicate set
+        n_partofdup = len(class_dups)
 
         class_dups.to_csv(duplicate_outfile)
         #print(class_dups.head(3))
 
+        # now throw away the duplicates (but keep the first of each set) from
+        # the main classifications table
         classifications = pd.DataFrame(classifications_nodups)
 
         del class_dups
         del is_a_dup
-        print("Duplicates removed from analysis.")
+        print("Duplicates removed from analysis (%d unique user-subject-workflow groups)." % n_dup_pairs)
 
     del subj_classifications
     del classifications_nodups
     gc.collect()
 
-
-# first, extract the started_at and finished_at from the metadata column
-classifications['meta_json'] = [json.loads(q) for q in classifications.metadata]
 
 classifications['created_day'] = [q[:10] for q in classifications.created_at]
 
@@ -293,10 +406,12 @@ last_class_day  = max(classifications.created_day).replace(' ', '')
 
 # save processing time and memory in the groupby.apply(); only keep the columns we're going to use or want to save
 if output_csv:
-    # if we'll be writing to a file at the end of this we need to save a few extra columns
-    cols_used = ["classification_id", "user_name", "user_id", "user_ip", "created_at", "created_day", "meta_json", "subject_ids", "workflow_id", "workflow_version"]
+    if not keep_allcols:
+        # if we'll be writing to a file at the end of this we need to save a few extra columns
+        cols_used = ["classification_id", "user_name", "user_id", "user_ip", "created_at", "created_day", "metadata", "meta_json", "subject_ids", "workflow_id", "workflow_version"]
 else:
-    cols_used = ["user_name", "user_id", "created_at", "created_day", "meta_json", "subject_ids"]
+    if not keep_allcols:
+        cols_used = ["classification_id", "user_name", "user_id", "user_ip", "created_at", "created_day", "meta_json", "subject_ids"]
 classifications = classifications[cols_used]
 # collect() calls PyInt_ClearFreeList(), so explicitly helps free some active memory
 gc.collect()
@@ -340,6 +455,10 @@ unregistered = [q.startswith("not-logged-in") for q in all_users]
 n_unreg = sum(unregistered)
 n_reg   = n_users_tot - n_unreg
 
+is_unreg_class = [q.startswith("not-logged-in") for q in classifications.user_name]
+n_unreg_class = sum(is_unreg_class)
+n_reg_class = n_class_tot - n_unreg_class
+
 # for the leaderboard, which I recommend project builders never make public because
 # Just Say No to gamification
 # But it's still interesting to see who your most prolific classifiers are, and
@@ -364,7 +483,8 @@ nclass_mean   = np.mean(nclass_byuser)
 nclass_gini   = gini(nclass_byuser)
 
 print("\nOverall:\n\n%d classifications of %d subjects by %d classifiers," % (n_class_tot,n_subj_tot,n_users_tot))
-print("%d logged in and %d not logged in, from %d unique IP addresses\n" % (n_reg,n_unreg,n_ip))
+print("%d logged in and %d not logged in, from %d unique IP addresses." % (n_reg,n_unreg,n_ip))
+print("%d classifications were from logged-in users, %d from not-logged-in users.\n" % (n_reg_class, n_unreg_class))
 print("That's %.2f classifications per subject on average (median = %.1f)." % (subj_class_mean, subj_class_med))
 print("The most classified subject has %d classifications; the least-classified subject has %d.\n" % (subj_class_max,subj_class_min))
 print("Median number of classifications per user: %.2f" %nclass_med)
@@ -447,21 +567,26 @@ if time_elapsed:
 
 if output_csv:
     # free up what memory we can before doing this (matters for big files)
-    del ok_times
-    del sa_temp
-    del fa_temp
+    if time_elapsed:
+        del ok_times
+        del sa_temp
+        del fa_temp
     del nclass_byuser
     del all_users
     del subj_class
     gc.collect()
 
-    classifications.to_csv(outfile_csv)
+    if keep_allcols:
+        classifications[cols_out].to_csv(outfile_csv)
+    else:
+        classifications.to_csv(outfile_csv)
     print("File with used subset of classification info written to %s ." % outfile_csv)
 
 print("File with ranked list of user classification counts written to %s ." % nclass_byuser_outfile)
 
-if remove_duplicates & (n_dups > 0):
-    print("Saved info for all classifications that have duplicates to %s ." % duplicate_outfile)
+if remove_duplicates:
+    if (n_dups > 0):
+        print("Saved info for all classifications that have duplicates to %s ." % duplicate_outfile)
 
 
 #end
